requirements.txt
supabase
python-dotenv
pandas
email
click
schedule
